Initial Commit
Now, that i decided to continue this project beyond the basic movement stage, version control is needed. To not be relying on unity too much git-for-untiy is used. Development currently has: Movement system, buildings: 2 and placeable, basic game mechanics like menus and the concept of a plan for a item system